+++ client/src/components/Header.js
@@ -1,4 +1,4 @@
-import { SignInModal } from './SignInModal';
+import { SignInModal } from '../pages/LoginPage/Components/SignInModal';
 import { Link, Outlet, useNavigate } from 'react-router-dom';
 import { loginStore, signInModalStore } from '../store/store';
 import { useState, useEffect } from 'react';
