+++ client/src/components/UserDropdown.js
@@ -5,7 +5,7 @@ import { Link, useNavigate } from 'react-router-dom';
 import { postLogout } from '../api/User';
 import { loginStore } from '../store/store';
 import { getUserId } from '../utils/cookies';
-import LogoutAlert from './LogoutAlert';
+import LogoutAlert from '../pages/LoginPage/Components/LogoutAlert';
 const UserDropdown = () => {
   const [anchorEl, setAnchorEl] = useState(null);
   const [open, setOpen] = useState(false);
