+++ client/src/pages/MainPage/Components/EmployAnnouncement.js
@@ -2,7 +2,8 @@ import ChevronLeftIcon from '@mui/icons-material/ChevronLeft';
 import ChevronRightIcon from '@mui/icons-material/ChevronRight';
 import { useEffect, useState } from 'react';
 import styled from 'styled-components';
-import { getEmployAnnouncement } from '../../api/Employ/get';
+// import { getEmployAnnouncement } from '../../api/Employ/get';
+import { getEmployAnnouncement } from '../../../api/Employ/get';
 const EmployAnnouncement = ({ mainCategory, subCategory }) => {
   const [employAnnouncement, setEmployAnnouncement] = useState(null);
   const [current, setCurrent] = useState(0);
