+++ client/src/pages/PostPage/Components/CommentEditForm.js
@@ -3,8 +3,8 @@ import MessageIcon from '@mui/icons-material/Message';
 import { useForm } from 'react-hook-form';
 import { useMemo, useEffect } from 'react';
 import styled from 'styled-components';
-import { putComment } from '../api/Comment';
-import { answerStore } from '../store/store';
+import { putComment } from '../../../api/Comment';
+import { answerStore } from '../../../store/store';
 import { useParams } from 'react-router-dom';
 const CommentEditForm = ({
   answer,
@@ -29,7 +29,7 @@ const CommentEditForm = ({
   useEffect(() => {
     reset(result);
   }, [result]);
-  const { question, getQuestion } = answerStore();
+  const { getQuestion } = answerStore();
   const commentEditSubmit = async (data) => {
     await putComment(data);
     await setSelectedComment(undefined);
