+++ client/src/pages/UserPage/Components/UserTab.js
@@ -2,10 +2,10 @@ import Box from '@mui/material/Box';
 import { useState } from 'react';
 import { Link, useNavigate, useParams } from 'react-router-dom';
 import styled from 'styled-components';
-import { deleteUser, postLogout } from '../../api/User';
-import { loginStore } from '../../store/store';
-import { getUserId } from '../../utils/cookies';
-import AlertDialog from '../AlertDialog';
+import { deleteUser, postLogout } from '../../../api/User';
+import { loginStore } from '../../../store/store';
+import { getUserId } from '../../../utils/cookies';
+import AlertDialog from '../../../components/AlertDialog';
 export default function UserTab({ menu, setMenu }) {
   const { loginHandler } = loginStore();
   const navigate = useNavigate();
