+++ client/src/pages/UserPage/ModifyUserPage.js
@@ -2,9 +2,9 @@ import styled from 'styled-components';
 import { Box, TextField, Grid, Container, Button } from '@mui/material';
 import { userStore } from '../../store/store';
 import { getUserId } from '../../utils/cookies';
-import ProfileModal from './ProfileModal';
+import ProfileModal from './Components/ProfileModal';
 import { useState, useEffect } from 'react';
-import UserPageAvatarWrapper from './UserPageAvatarWrapper';
+import UserPageAvatarWrapper from './Components/UserPageAvatarWrapper';
 import CreateIcon from '@mui/icons-material/Create';
 import { modifyPassword, modifyUser } from '../../api/User';
 import { useForm } from 'react-hook-form';
