--- client/src/utils/delay.js
@@ -1,3 +1,0 @@
-const delay = (ms) => new Promise((resolve) => setTimeout(resolve, ms));
-
-export default delay;
